Agregar permisos a controladores

+++ app/Http/Controllers/NotificationController.php
@@ -6,10 +6,15 @@ use Illuminate\Http\Request;
 
 use Auth;
 use DB;
+use Gate;
 
 class NotificationController extends Controller
 {
     public function show() {
+
+        if (Gate::denies('show-notifications')) {
+            abort(403);
+        }
 
     	$user = Auth::User();
 
@@ -19,6 +24,10 @@ class NotificationController extends Controller
     }
 
     public function destroy() {
+
+        if (Gate::denies('delete-notifications')) {
+            abort(403);
+        }
 
         DB::delete('DELETE FROM fashionrecovery."GR_040" WHERE "UserID"='.Auth::User()->id);
